fix: signed/unsigned comparison warnings (-Wsign-compare)
Fix two HIGH priority signed/unsigned comparison warnings:

1. MySQL_HostGroups_Manager.cpp:1792 - Change loop variable from
   'int j' to 'unsigned int j' to match the unsigned comparison
   with servers->len.

2. MySQL_Thread.cpp:4357 - Cast int to size_t when comparing with
   getBufferSize() which returns size_t.

These changes prevent potential logic errors from negative values
being interpreted as large unsigned values.

Found during compiler warning analysis with 'make debug'.

+++ lib/MySQL_Thread.cpp
@@ -4354,7 +4354,7 @@ void MySQL_Thread::refresh_variables() {
 	REFRESH_VARIABLE_INT(eventslog_buffer_history_size);
 	{
 		int elmhs = mysql_thread___eventslog_buffer_history_size;
-		if (GloMyLogger->MyLogCB->getBufferSize() != elmhs) {
+		if (GloMyLogger->MyLogCB->getBufferSize() != static_cast<size_t>(elmhs)) {
 			GloMyLogger->MyLogCB->setBufferSize(elmhs);
 		}
 	}
